% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformation_fat_flp.R
\name{build_freq_conf_fat}
\alias{build_freq_conf_fat}
\title{Build carcass conformation and fat frequencies depending on slaughtercategory}
\usage{
build_freq_conf_fat(
  ps_input_flp_tibble,
  ps_sex,
  ps_marketing_channel,
  ps_flag_cow,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_input_flp_tibble}{input flp tibble coming from read_file_input_flp in this package}

\item{ps_sex}{statement of sex ("F" for female and "M" for male)}

\item{ps_marketing_channel}{statement of marketing-channel for Natura-Beef (==2), SwissPrimBeef(==3)}

\item{ps_flag_cow}{flag to select for cows or not (TRUE or FALSE)}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
result_matrix matrix with frequencies for carcass conformation and fat
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will prepare carcass conformation and fat matrices
depending on slaughtercategory.
}
