% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calving.R
\name{calculate_abortion_rate}
\alias{calculate_abortion_rate}
\title{Calculate abortion rate}
\usage{
calculate_abortion_rate(
  ps_input_calving_tibble,
  ps_statement_firstlactation = TRUE,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_calving_tibble}{input calving tibble coming from read_file_input_calving in this package}

\item{ps_statement_firstlactation}{statement if in first lactation status (TRUE or FALSE)}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
abortion_rate vector
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate based on calving scores as well
as lactation number of the dam (primiparous vs multiparous) the abortion rate.
}
