% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_data_flp.R
\name{calculate_age_calf_sale}
\alias{calculate_age_calf_sale}
\title{calculate_age_calf_sale}
\usage{
calculate_age_calf_sale(
  ps_input_flp_tibble,
  ps_sex,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_flp_tibble}{input flp tibble coming from read_file_input_flp in this package}

\item{ps_sex}{sex of the animal}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
rearing_age
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate rearing weight of calves.
}
