% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_data_flp.R
\name{calculate_bull_liveweight}
\alias{calculate_bull_liveweight}
\title{Calculate bull mature live weight}
\usage{
calculate_bull_liveweight(
  ps_input_flp_tibble,
  ps_sirebreed,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_flp_tibble}{input flp tibble coming from read_file_input_flp in this package}

\item{ps_sirebreed}{sire breed}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
bull_liveweight vector
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate bull live weight.
}
