% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calving.R
\name{calculate_calvesdied24h_proportion}
\alias{calculate_calvesdied24h_proportion}
\title{Calculate proportion of calves died to 24 hours}
\usage{
calculate_calvesdied24h_proportion(
  ps_input_calving_tibble,
  ps_statement_firstlactation = TRUE,
  ps_statement_easycalving = TRUE,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_calving_tibble}{input calving tibble coming from read_file_input_calving in this package}

\item{ps_statement_firstlactation}{statement if in first lactation status (TRUE or FALSE)}

\item{ps_statement_easycalving}{statement if for easy calving status (TRUE or FALSE)}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
calves died to 24 hours proportion vector
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate based on calving score,
lactation number of the dam (primiparous vs multiparous) to calculate the proportion of calves died.
The parameter in ECOWEIGHT specifies within 48 hours after birth. However, we only have data specified for up to 24hours after birth.
Therefore we use 24hours in this function.
}
