% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calving.R
\name{calculate_calvesdiedafter24h_proportion}
\alias{calculate_calvesdiedafter24h_proportion}
\title{Calculate proportion of calves died after 24 hours}
\usage{
calculate_calvesdiedafter24h_proportion(
  ps_input_calving_tibble,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_calving_tibble}{input calving tibble coming from read_file_input_calving in this package}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
calves died after 24 hours proportion vector
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate the proportion of calves died after 24h.
It would be nice to have the information after 48h to weaning. But at this stage, we don't.
}
