% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calving.R
\name{calculate_calvingscore_proportion}
\alias{calculate_calvingscore_proportion}
\title{Calculate calving score proportions}
\usage{
calculate_calvingscore_proportion(
  ps_input_calving_tibble,
  ps_statement_firstlactation = TRUE,
  ps_breed,
  ps_sex,
  ps_calvingscore,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_calving_tibble}{input calving tibble coming from read_file_input_calving in this package}

\item{ps_statement_firstlactation}{statement if in first lactation status (TRUE or FALSE)}

\item{ps_breed}{set the breed (AN, AU, CH, LM, SI, OB)}

\item{ps_sex}{set the sex (F, M)}

\item{ps_calvingscore}{set the calving score for which the proportion should be calculated (2,3,4)}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
calving score proportion vector
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate based on breed, sex as well
as lactation number of the dam (primiparous vs multiparous) the calving score proportion.
}
