% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_data_flp.R
\name{calculate_carcass_price}
\alias{calculate_carcass_price}
\title{calculate_carcass_price}
\usage{
calculate_carcass_price(
  ps_tbl_input_statement_carcass,
  ps_sex,
  ps_liveweight,
  ps_marketchannel,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_sex}{category e.g MT (bull), RG(heifer) or KV(veal calf)}

\item{ps_liveweight}{value for liveweight}

\item{ps_marketchannel}{marketing chanell to determine conventional beef or veal}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}

\item{ps_tbl_input_statemement_carcass}{tibble of input statements including base carcass prices}
}
\value{
rearing_age
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate rearing weight of calves.
}
