% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_data_flp.R
\name{calculate_cow_liveweight}
\alias{calculate_cow_liveweight}
\title{Calculate cow live weight}
\usage{
calculate_cow_liveweight(
  ps_input_flp_tibble,
  ps_first_calvingweight,
  ps_second_calvingweight,
  ps_dambreed,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_flp_tibble}{input flp tibble coming from read_file_input_flp in this package}

\item{ps_first_calvingweight}{flag to calculate first calving weight (TRUE or FALSE)}

\item{ps_second_calvingweight}{flag to calculate second calving weight (TRUE or FALSE)}

\item{ps_dambreed}{dam breed}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
cow_liveweight vector
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate cow live weight
}
