% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_data_flp.R
\name{calculate_dailygain}
\alias{calculate_dailygain}
\title{Calculate average daily gain}
\usage{
calculate_dailygain(
  pv_mean_slaughterage,
  pv_mean_weaningage,
  pv_mean_livewt_atslaughter,
  pv_mean_weaningwt,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{pv_mean_slaughterage}{vector with the mean of slaughter age}

\item{pv_mean_weaningage}{vector with the mean of weaning age}

\item{pv_mean_livewt_atslaughter}{vector with the mean of live weight at slaughter}

\item{pv_mean_weaningwt}{vector with the mean of weaning weight}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
dailygain vector
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate average daily gain.
}
