% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_data_flp.R
\name{calculate_dailygain_rearing}
\alias{calculate_dailygain_rearing}
\title{Calculate average daily gain during rearing}
\usage{
calculate_dailygain_rearing(
  pv_mean_rearingage,
  pv_mean_rearing_wt,
  pv_mean_birthwt,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{pv_mean_rearingage}{vector with the mean of slaughter age}

\item{pv_mean_rearing_wt}{vector with the mean of live weight at slaughter}

\item{pv_mean_birthwt}{vector with the mean of weaning weight}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
dailygain_rearing vector
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate average daily gain.
}
