% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_data_flp.R
\name{calculate_extrapolated_weaningweight}
\alias{calculate_extrapolated_weaningweight}
\title{Extrapolate weaning weight at t-days}
\usage{
calculate_extrapolated_weaningweight(
  pv_mean_weaningage,
  pv_daily_gain,
  pv_mean_weaningwt,
  pv_t_days,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{pv_mean_weaningage}{vector with the mean of weaning age}

\item{pv_daily_gain}{vector with daily gain}

\item{pv_mean_weaningwt}{vector with the mean of weaning weight}

\item{pv_t_days}{vector for t-days}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
weaningweight_tdays vector
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate the extrapolate weaning weight
at t-days.
}
