% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_data_flp.R
\name{calculate_mean_weaningweight}
\alias{calculate_mean_weaningweight}
\title{Calculate mean weaning weight}
\usage{
calculate_mean_weaningweight(
  ps_input_flp_tibble,
  ps_sex,
  ps_marketing_channel,
  ps_prodsystew,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_flp_tibble}{input flp tibble coming from read_file_input_flp in this package}

\item{ps_sex}{statement of sex ("F" for female and "M" for male)}

\item{ps_marketing_channel}{statement of marketing-channel for Natura-Beef (==2), SwissPrimBeef(==3)}

\item{ps_prodsystew}{production system build up as option in ECOWEIGHT}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
weaningwt vector
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate mean weaning weight
based on slaughtercategory, marketing-channel (label of Swiss Beef Cattle Assiocation) and sex.
}
