% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/progeny_data_flp.R
\name{calculate_rearing_age_beef}
\alias{calculate_rearing_age_beef}
\title{calculate_rearing_age_beef}
\usage{
calculate_rearing_age_beef(ps_input_flp_tibble, pb_log = FALSE, plogger = NULL)
}
\arguments{
\item{ps_input_flp_tibble}{input flp tibble coming from read_file_input_flp in this package}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
rearing_age
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate rearing age of beef animals.
}
