% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calving.R
\name{calculate_stillbirth_rate}
\alias{calculate_stillbirth_rate}
\title{Calculate stillbirth rate}
\usage{
calculate_stillbirth_rate(
  ps_input_calving_tibble,
  ps_statement_firstlactation = TRUE,
  ps_statement_easycalving = TRUE,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_calving_tibble}{input calving tibble coming from read_file_input_calving in this package}

\item{ps_statement_firstlactation}{statement if in first lactation status (TRUE or FALSE)}

\item{ps_statement_easycalving}{statement if for easy calving status (TRUE or FALSE)}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
stillbirth_rate vector
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will calculate based on calving scores as well
as lactation number of the dam (primiparous vs multiparous) the stillbirth rate.
}
