% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{calculate_transformed_EW_ewdc}
\alias{calculate_transformed_EW_ewdc}
\title{Calculate the economic weight for calving score on the transformed scale used for ebv}
\usage{
calculate_transformed_EW_ewdc(
  ps_input_file_calving,
  ps_start_calving_date,
  ps_end_calving_data,
  ps_sirebreed,
  ps_dambreed,
  ps_EW_calving,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_input_file_calving}{calving data used for the calculation of calving score proportions}

\item{ps_start_calving_date}{start date for calving data}

\item{ps_end_calving_data}{end date for calving data}

\item{ps_EW_calving}{economic weight for calving score}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}

\item{ps_sire_breed}{sire breed}

\item{ps_dam_breed}{dam breed}
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function processed different functions
to prepare information to be plot.
}
