% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/change_input_parameter.R
\name{change_input_paramfile}
\alias{change_input_paramfile}
\title{Change input-parameter-file for ECOWEIGHT}
\usage{
change_input_paramfile(
  ps_sirebreed,
  ps_dambreed,
  ps_prodsystew,
  ps_marketchannel,
  ps_path_input_paramfile,
  ps_name_input_paramfile_2change,
  ps_name_input_paramfile_changed,
  ps_statement2search,
  ps_value2change,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_sirebreed}{sire breed}

\item{ps_dambreed}{dam breed}

\item{ps_prodsystew}{production system build up as option in ECOWEIGHT}

\item{ps_marketchannel}{market channel}

\item{ps_name_input_paramfile_2change}{path to input-parameter-file to change for ECOWEIGHT}

\item{ps_name_input_paramfile_changed}{path to input-parameter-file changed for ECOWEIGHT}

\item{ps_statement2search}{statement to search}

\item{ps_value2change}{how to change the value}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}

\item{ps_path_directory2create}{path of the directory that will be created}
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will change the input parameter.
}
