% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{combine_popmean}
\alias{combine_popmean}
\title{Combination of the results from the output-parameter-file of ECOWEIGHT}
\usage{
combine_popmean(
  ptbl_popmean_results,
  ptbl_EW_results,
  ps_scenario,
  ps_prodsystem,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ptbl_popmean_results}{tibble with population mean out of ECOWEIGHT}

\item{ptbl_EW_results}{tibble with economic weight out of ECOWEIGHT}

\item{ps_scenario}{string for the scenario}

\item{ps_prodsystem}{production system build up as option in ECOWEIGHT}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
tbl_aggregate_results result tibble with economic weight and population mean
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function combine economic weight and population mean.
}
