% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{combine_popmean_ewbc}
\alias{combine_popmean_ewbc}
\title{Combination of the results from the output-parameter-file of ECOWEIGHT beef on beef}
\usage{
combine_popmean_ewbc(
  ptbl_popmean_results,
  ptbl_EW_results,
  ps_scenario,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ptbl_popmean_results}{tibble with population mean out of ECOWEIGHT}

\item{ptbl_EW_results}{tibble with economic weight out of ECOWEIGHT}

\item{ps_scenario}{string for the scenario}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
tbl_aggregate_results result tibble with economic weight and population mean
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function combine economic weight and population mean.
}
