% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_scenario.R
\name{create_directory_scenario}
\alias{create_directory_scenario}
\title{Create directory per scenario with input-parameter-file for ECOWEIGHT}
\usage{
create_directory_scenario(
  ps_sirebreed,
  ps_dambreed,
  ps_prodsystew,
  ps_marketchannel,
  ps_path_directory2create,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_sirebreed}{sire breed}

\item{ps_dambreed}{dam breed}

\item{ps_prodsystew}{production system build up as option in ECOWEIGHT}

\item{ps_marketchannel}{market channel}

\item{ps_path_directory2create}{path of the directory that will be created}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will create a directory with a specific name
to recognize each scenario and file this directory with required input-parameter-file
to run ECOWEIGHT.
}
