% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{create_table_results_ewbc}
\alias{create_table_results_ewbc}
\title{Create table of results depending on sire breed, dam breed or production system}
\usage{
create_table_results_ewbc(
  ps_sort_by,
  ps_path_results_tbl,
  ps_path_save,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_sort_by}{parameter to determine how to sort the tables: sire_breed, dam_breed or production_system}

\item{ps_path_results_tbl}{path to results tables from initial post processing}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}

\item{ps_path_tbl_save}{path to save the results: cannot be the same path as the results tables from inital post processing}
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function processed different functions
to prepare information to be plot.
}
