% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{extract_avg_fattening_length}
\alias{extract_avg_fattening_length}
\title{Extract average fattening length from the output-parameter-file of ECOWEIGHT}
\usage{
extract_avg_fattening_length(
  ps_path_2outputfile,
  ps_tbl_output_statement,
  ps_tbl_output_search_pattern,
  ps_prodsystem,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_path_2outputfile}{path to output file of ECOWEIGHT}

\item{ps_tbl_output_statement}{tibble with output statement}

\item{ps_prodsystem}{production system build up as option in ECOWEIGHT}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}

\item{ps_output_search_pattern}{output file with the search patterns}
}
\value{
avg_fattening_length output average fattening length
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function extract the average fattening length.
}
