% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{extract_birth_wean_weight_popmean_ewbc}
\alias{extract_birth_wean_weight_popmean_ewbc}
\title{Extract birth and weaning weight from the output-parameter-file of ECOWEIGHT beef on beef}
\usage{
extract_birth_wean_weight_popmean_ewbc(
  ps_path_2outputfile,
  ps_tbl_output_statement,
  ps_tbl_output_search_pattern,
  pl_constants_postprocess_beefOnbeef,
  ps_marketchannel,
  ps_prodsystew,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_path_2outputfile}{path to output file of ECOWEIGHT}

\item{ps_tbl_output_statement}{tibble with output statement}

\item{ps_tbl_output_search_pattern}{tibble with the search patterns}

\item{pl_constants_postprocess_beefOnbeef}{constants of postprocessing beefOnbeef}

\item{ps_marketchannel}{market channel}

\item{ps_prodsystew}{production system build up as option in ECOWEIGHT}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
tbl_result_birth_wean tibble results with birth and weaning weight
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function extract the birth and weaning weight of the population mean.
}
