% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{extract_calv_flesh_fat_popmean_ewbc}
\alias{extract_calv_flesh_fat_popmean_ewbc}
\title{Extract calving score, fleshiness and fat from the output-parameter-file of ECOWEIGHT beef on beef}
\usage{
extract_calv_flesh_fat_popmean_ewbc(
  ps_path_2outputfile,
  ps_tbl_output_statement,
  ps_tbl_output_search_pattern,
  pl_constants_postprocess_beefOnbeef,
  ps_marketchannel,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_path_2outputfile}{path to output file of ECOWEIGHT}

\item{ps_tbl_output_statement}{tibble with output statement}

\item{ps_tbl_output_search_pattern}{tibble with the search patterns}

\item{pl_constants_postprocess_beefOnbeef}{constants of postprocessing beefOnbeef}

\item{ps_marketchannel}{market channel}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
tbl_result_misc tibble results with calving, fleshiness, fat
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function extract the calving score, fleshiness, fat of the population mean.
}
