% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{extract_ewbc}
\alias{extract_ewbc}
\title{Extract economic weights from the output-parameter-file of ECOWEIGHT beef on beef}
\usage{
extract_ewbc(
  ps_path_2outputfile,
  ps_output_statement,
  ps_output_search_pattern,
  ps_sirebreed,
  ps_dambreed,
  ps_prodsystew,
  ps_marketchannel,
  ps_path_directory2create,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_path_2outputfile}{path to output file of ECOWEIGHT}

\item{ps_output_statement}{output statement in a file}

\item{ps_output_search_pattern}{output file with the search patterns}

\item{ps_sirebreed}{sire breed}

\item{ps_dambreed}{dam breed}

\item{ps_prodsystew}{production system build up as option in ECOWEIGHT}

\item{ps_marketchannel}{market channel}

\item{ps_path_directory2create}{path of the directory that will be created}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
tbl_aggregate_results result tibble with economic weights and units
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function extract the economic weights.
}
