% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{extract_gestlength_ecovalue_ewdc}
\alias{extract_gestlength_ecovalue_ewdc}
\title{Extract gestation length economic value from the output-parameter-file of ECOWEIGHT beef on dairy}
\usage{
extract_gestlength_ecovalue_ewdc(
  ps_path_2outputfile_run0,
  ps_path_2outputfile_run1,
  ps_tbl_output_statement,
  ps_tbl_output_search_pattern,
  pl_constants_postprocess_beefOndairy,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_path_2outputfile_run0}{path to output file run 0 of ECOWEIGHT}

\item{ps_path_2outputfile_run1}{path to output file run 1 of ECOWEIGHT}

\item{ps_tbl_output_statement}{tibble with output statement}

\item{ps_tbl_output_search_pattern}{tibble with the search patterns}

\item{pl_constants_postprocess_beefOndairy}{constants of postprocessing beefOndairy}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
tbl_profit_gestlength tibble results with gestation length economic value
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function extract gestation length economic value.
}
