% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{extract_gestlength_popmean_ewdc}
\alias{extract_gestlength_popmean_ewdc}
\title{Extract gestation length from the output-parameter-file of ECOWEIGHT beef on dairy}
\usage{
extract_gestlength_popmean_ewdc(
  ps_path_2outputfile,
  ps_tbl_output_statement,
  ps_tbl_output_search_pattern,
  pl_constants_postprocess_beefOndairy,
  pl_constants_progeny_beefOndairy,
  ps_marketchannel,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_path_2outputfile}{path to output file of ECOWEIGHT}

\item{ps_tbl_output_statement}{tibble with output statement}

\item{ps_tbl_output_search_pattern}{tibble with the search patterns}

\item{pl_constants_postprocess_beefOndairy}{constants of postprocessing beefOndairy}

\item{pl_constants_progeny_beefOndairy}{constants of progeny beefOndairy}

\item{ps_marketchannel}{market channel}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
tbl_result_gestation tibble results with gestation length population mean
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function extract gestation length population mean.
}
