% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results.R
\name{extract_result}
\alias{extract_result}
\title{Extract the part of interest of the results coming from ECOWEIGHT output}
\usage{
extract_result(
  ps_path_2outputfile,
  ps_start_statement2extract,
  ps_end_statement2extract,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_path_2outputfile}{path to output file of ECOWEIGHT}

\item{ps_start_statement2extract}{string as statement to set the beginning of the extraction}

\item{ps_end_statement2extract}{string as statement to set the ending of the extraction}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
vec_ecow_result_2extract
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce ouptut file with the results. This function will extract the results needed.
}
