% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_constants_calving_beefOndairy}
\alias{get_constants_calving_beefOndairy}
\title{Get constants for calving in beef-on-dairy production systems}
\usage{
get_constants_calving_beefOndairy()
}
\description{
Return a list with specific calving constants in beef-on-beef production systems
that are used.
}
