% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_constants_carcass_beefOnbeef}
\alias{get_constants_carcass_beefOnbeef}
\title{Get constants for carcass in beef-on-beef production systems}
\usage{
get_constants_carcass_beefOnbeef()
}
\description{
Return a list with specific constants that are used.
}
