% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_constants_ew_input_beefOndairy}
\alias{get_constants_ew_input_beefOndairy}
\title{Get constants for ew input in beef-on-dairy production systems}
\usage{
get_constants_ew_input_beefOndairy()
}
\description{
Return a list with constants in beef-on-dairy production systems
that are used.
}
