% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_constants_ewbc_input_beefOnbeef}
\alias{get_constants_ewbc_input_beefOnbeef}
\title{Get constants for ewbc input in beef-on-beef production systems}
\usage{
get_constants_ewbc_input_beefOnbeef()
}
\description{
Return a list with specific calving constants in beef-on-beef production systems
that are used.
}
