% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_constants_liveweight_deductions_beefOnbeef}
\alias{get_constants_liveweight_deductions_beefOnbeef}
\title{Get constants for carcass weight deductions based on liveweight for NaturaBeef}
\usage{
get_constants_liveweight_deductions_beefOnbeef()
}
\description{
Return a list with specific constants that are used.
}
