% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_constants_liveweight_deductions_veal_beefOndairy}
\alias{get_constants_liveweight_deductions_veal_beefOndairy}
\title{Get constants for carcass weight deductions based on liveweight for veal}
\usage{
get_constants_liveweight_deductions_veal_beefOndairy()
}
\description{
Return a list with specific constants that are used.
}
