% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_constants_postprocess_beefOnbeef}
\alias{get_constants_postprocess_beefOnbeef}
\title{Get constants for post processing for beef-on_beef}
\usage{
get_constants_postprocess_beefOnbeef()
}
\description{
Return a list with specific constants that are used.
}
