% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_constants_postprocess_beefOndairy}
\alias{get_constants_postprocess_beefOndairy}
\title{Get constants for post processing for beef-on_dairy}
\usage{
get_constants_postprocess_beefOndairy()
}
\description{
Return a list with specific constants that are used.
}
