% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_constants_progeny_beefOndairy}
\alias{get_constants_progeny_beefOndairy}
\title{Get constants for progeny in beef-on-dairy production systems}
\usage{
get_constants_progeny_beefOndairy()
}
\description{
Return a list with specific constants that are used.
}
