% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{get_constants_transform_fleshiness}
\alias{get_constants_transform_fleshiness}
\title{Get constants for the transformation of fleshiness}
\usage{
get_constants_transform_fleshiness()
}
\description{
Return a list with constants that are used.
}
