% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{get_qp4ewc_logger}
\alias{get_qp4ewc_logger}
\title{Create log4r Logger for package}
\usage{
get_qp4ewc_logger(ps_logfile = "qp4ewc.log", ps_level = "FATAL")
}
\arguments{
\item{ps_logfile}{name of the logfile}

\item{ps_level}{logger level}
}
\value{
qp4ewc_logger
}
\description{
Create log4r Logger for package
}
