% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{plot_piechart_ewdc}
\alias{plot_piechart_ewdc}
\title{Plot pie chart of the results coming from ECOWEIGHT dairy beef cattle}
\usage{
plot_piechart_ewdc(
  ps_path_2genSD,
  ptbl_EW_results,
  ps_traitgroup2consider,
  ps_scenario,
  ps_marketchannel,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_path_2genSD}{path to file with genetic standard deviation}

\item{ptbl_EW_results}{tibble of economic weights}

\item{ps_traitgroup2consider}{traitgroup may be Carcass or Functional Traits}

\item{ps_scenario}{name of the scenario (includes sire breed, dam breed, production system and marketing channel)}

\item{ps_marketchannel}{marketing channel required for determining the correct genetic standard deviation to use}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function processed different functions
to prepare information to be plot.
}
