% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{post_process_ewbc_output}
\alias{post_process_ewbc_output}
\title{Post-processing the output-parameter-file of ECOWEIGHT beef cattle}
\usage{
post_process_ewbc_output(
  ps_path_2outputfile,
  ps_output_statement,
  ps_output_search_pattern,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_path_2outputfile}{path to output file of ECOWEIGHT}

\item{ps_output_statement}{output statement in a file}

\item{ps_output_search_pattern}{output file with the search patterns}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function processed different functions
to extract information from the output parameter files.
}
