% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{pre_process_ew_input_calving}
\alias{pre_process_ew_input_calving}
\title{Pre-processing the calving data for input-parameter-file of ECOWEIGHT}
\usage{
pre_process_ew_input_calving(
  ps_sirebreed,
  ps_prodsystew,
  ps_marketchannel,
  ps_path_directory2create,
  ps_input_file_calving_statement,
  ps_input_file_calving,
  ps_start_calving_date,
  ps_end_calving_date,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_sirebreed}{sire breed}

\item{ps_prodsystew}{production system build up as option in ECOWEIGHT}

\item{ps_marketchannel}{market channel}

\item{ps_path_directory2create}{path of the directory that will be created}

\item{ps_input_file_calving_statement}{path to file with statement based on calving for the input-parameter-file for ECOWEIGHT}

\item{ps_input_file_calving}{path to file with input coming from calving for the input-parameter-file for ECOWEIGHT}

\item{ps_start_calving_date}{setting the start of the calving date to filter in the calving data}

\item{ps_end_calving_date}{setting the end of the calving date to filter in the calving data}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function processed different functions
to prepare the input parameter files based on calving data.
}
