% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{pre_process_ewdc_input_carcass_data_flp}
\alias{pre_process_ewdc_input_carcass_data_flp}
\title{Pre-processing the carcass conformation, fat, prices based on flp-data for beef-on-dairy input-parameter-file of ECOWEIGHT}
\usage{
pre_process_ewdc_input_carcass_data_flp(
  ps_sirebreed,
  ps_dambreed,
  ps_prodsystew,
  ps_marketchannel,
  ps_path_directory2create,
  ps_input_file_flp_carcass_matrix_statement,
  ps_input_file_flp,
  ps_input_file_ped,
  ps_start_flp_date,
  ps_end_flp_date,
  ps_input_file_price_cow,
  ps_input_file_price_bull,
  ps_input_file_price_heifer,
  ps_input_file_price_calf,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_sirebreed}{sire breed}

\item{ps_dambreed}{dam breed}

\item{ps_prodsystew}{production system build up as option in ECOWEIGHT}

\item{ps_marketchannel}{market channel}

\item{ps_path_directory2create}{path of the directory that will be created}

\item{ps_input_file_flp_carcass_matrix_statement}{path to file with statement based on calving for the input-parameter-file for ECOWEIGHT}

\item{ps_input_file_flp}{path to file with input coming from calving for the input-parameter-file for ECOWEIGHT}

\item{ps_input_file_ped}{input file for pedigree}

\item{ps_start_flp_date}{setting the start of the slaughter date to filter in the slaughter data pre_process_input_progeny_data_flp}

\item{ps_end_flp_date}{setting the end of the slaughter date to filter in the slaughter data}

\item{ps_input_file_price_cow}{path to file with price for cows}

\item{ps_input_file_price_bull}{path to file with price for bulls}

\item{ps_input_file_price_heifer}{path tho file with price for heifers}

\item{ps_input_file_price_calf}{path tho file with price for heifers}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function processed different functions
to prepare the input parameter files based on progeny flp data and literature about the price-system.
}
