% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{qp4ewc_log_info}
\alias{qp4ewc_log_info}
\title{Wrapper for log4r info}
\usage{
qp4ewc_log_info(plogger, ps_caller, ps_msg)
}
\arguments{
\item{plogger}{log4r logger object}

\item{ps_msg}{logging message}
}
\description{
Wrapper for log4r info
}
