% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_file_input}
\alias{read_file_input}
\title{Read file with input for input-parameter-file of ECOWEIGHT}
\usage{
read_file_input(ps_input_file, pb_log = FALSE, plogger = NULL)
}
\arguments{
\item{ps_input_file}{path to file with input for the input-parameter-file for ECOWEIGHT}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
tibble with the content of the file
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will read a file with value .
}
