% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_file_input_calf}
\alias{read_file_input_calf}
\title{Read file with pedigree input}
\usage{
read_file_input_calf(ps_input_file_calf, pb_log = FALSE, plogger = NULL)
}
\arguments{
\item{ps_input_file_calf}{path to calf price file}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
tibble with the pedigree content
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will read a pedigree file.
}
