% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_file_input_calving}
\alias{read_file_input_calving}
\title{Read file with input about calving for input-parameter-file of ECOWEIGHT}
\usage{
read_file_input_calving(
  ps_input_file_calving,
  ps_start_calving_date,
  ps_end_calving_date,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_file_calving}{path to file with input coming from calving for the input-parameter-file for ECOWEIGHT}

\item{ps_start_calving_date}{setting the start of the calving date to filter in the calving data}

\item{ps_end_calving_date}{setting the end of the calving date to filter in the calving data}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
tibble with the content of the calving file
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will read a file with value coming from calving.
In this calving file some information about abortion rate, stillbirth rate, calving score,
amount of calves died within 48 hours after birth.
}
