% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_file_input_flp}
\alias{read_file_input_flp}
\title{Read file with input about beef recording for input-parameter-file of ECOWEIGHT}
\usage{
read_file_input_flp(
  ps_input_file_flp,
  ps_start_flp_date,
  ps_end_flp_date,
  ps_sirebreed,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_file_flp}{path to file with input coming from beef recording for the input-parameter-file for ECOWEIGHT}

\item{ps_start_flp_date}{setting the start of the beef recording date to filter in the flp-data}

\item{ps_end_flp_date}{setting the end of the beef recording date to filter in the flp-data}

\item{ps_sirebreed}{sire breed}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
tibble with the content of the beef recording file
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will read a file with value coming from weights.
In this weight-file some information about ...
}
