% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_file_input_literature}
\alias{read_file_input_literature}
\title{Read file with input from literature research for input-parameter-file of ECOWEIGHT}
\usage{
read_file_input_literature(
  ps_input_file_literature,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_file_literature}{path to file with input coming from literature for the input-parameter-file for ECOWEIGHT}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
tibble with the content of the literatur file
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will read a file with value coming from literature research.
}
