% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.R
\name{read_price_conf_fat}
\alias{read_price_conf_fat}
\title{Read prices of a specific slaughtercategory}
\usage{
read_price_conf_fat(ps_input_file_price, pb_log = FALSE, plogger = NULL)
}
\arguments{
\item{ps_input_file_price}{path to file with price for a specific slaughtercategory}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
result_price_matrix matrix with prices for carcass conformation and fat
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will read the price for a slaughtercategory depending of
carcass conformation and fat.
}
