% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{save_csv_table_piechart_ewdc}
\alias{save_csv_table_piechart_ewdc}
\title{Save table and piechart in pdf}
\usage{
save_csv_table_piechart_ewdc(
  ps_tbl_aggregate_results,
  ptbl_EW_results,
  ps_path_tbl_save,
  ps_scenario,
  ps_marketchannel,
  ps_input_genetic_SD,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_tbl_aggregate_results}{tibble with economic weight and population mean}

\item{ptbl_EW_results}{tibble of economic weights}

\item{ps_path_tbl_save}{path to save results}

\item{ps_scenario}{name of the scenario (includes sire breed, dam breed, production system and marketing channel)}

\item{ps_marketchannel}{marketing channel required for determining the correct genetic standard deviation to use}

\item{ps_input_genetic_SD}{input file with genetic standarddeviation}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function process the results in a pdf.
}
