% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conformation_fat_flp.R
\name{transform_fleshinesss}
\alias{transform_fleshinesss}
\title{Transformation of the fleshiness score}
\usage{
transform_fleshinesss(ps_input_merged_tibble, pb_log, plogger = NULL)
}
\arguments{
\item{ps_input_merged_tibble}{input merged tibble with flp and pedigree data}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\value{
tbl_merged_data tibble with a further column containing transformed fleshiness score
}
\description{
The raw fleshiness score need to be transformed like in the genetic evaluation.
}
