% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updating.R
\name{update_input_parameter_file}
\alias{update_input_parameter_file}
\title{Update input-parameter-file of ECOWEIGHT}
\usage{
update_input_parameter_file(
  ps_path2template_input_parameter_file,
  ps_statement2search,
  ps_value2update,
  ps_line4statement2update = 1,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_path2template_input_parameter_file}{path to the template of the input-parameter-file for ECOWEIGHT}

\item{ps_statement2search}{statement to search in the template}

\item{ps_value2update}{value to update in the input-parameter-file}

\item{ps_line4statement2update}{numeric value of the number of the line before the statement to update}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function will search a specific statement in a template and
will update the value for the input parameter files.
}
