% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calving.R
\name{write_mean_sd_rawAndtransform_calvscore}
\alias{write_mean_sd_rawAndtransform_calvscore}
\title{Write mean and standarddeviation for raw and transformed calving score}
\usage{
write_mean_sd_rawAndtransform_calvscore(
  ps_input_tbl_calving = tbl_calving,
  ps_sirebreed = ps_sirebreed,
  ps_dambreed = ps_dambreed,
  ps_prodsystew = ps_prodsystew,
  ps_marketchannel = ps_marketchannel,
  ps_path_directory2create = ps_path_directory2create,
  pb_log = FALSE,
  plogger = NULL
)
}
\arguments{
\item{ps_input_tbl_calving}{input calving tibble coming from read_file_input_calving in this package}

\item{ps_sirebreed}{sire breed}

\item{ps_dambreed}{dam breed}

\item{ps_prodsystew}{production system build up as option in ECOWEIGHT}

\item{ps_marketchannel}{market channel}

\item{ps_path_directory2create}{path of the directory that will be created}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\description{
Raw calving scores are not in the same scale for the genetic evaluation.
Therefore some further information like the mean and standarddeviation need
to be calculated for raw as well as for transformed calving score scale,
to insure to use the calculated economic weight out of ECOWEIGHT.
}
