% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_processing.R
\name{post_process_ewbc_output}
\alias{post_process_ewbc_output}
\title{Post-processing the output-parameter-file of ECOWEIGHT beef cattle}
\usage{
post_process_ewbc_output(
  ps_path_2outputfile,
  ps_output_statement,
  ps_output_search_pattern,
  ps_sirebreed,
  ps_dambreed,
  ps_prodsystew,
  ps_marketchannel,
  ps_path_directory2create,
  ps_scenario,
  ps_path_tbl_save,
  ps_input_genetic_SD,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_path_2outputfile}{path to output file of ECOWEIGHT}

\item{ps_output_statement}{output statement in a file}

\item{ps_output_search_pattern}{output file with the search patterns}

\item{ps_sirebreed}{sire breed}

\item{ps_dambreed}{dam breed}

\item{ps_prodsystew}{production system build up as option in ECOWEIGHT}

\item{ps_marketchannel}{market channel}

\item{ps_path_directory2create}{path of the directory that will be created}

\item{ps_scenario}{name of the scenario (includes sire breed, dam breed, production system and marketing channel)}

\item{ps_path_tbl_save}{path to save results}

\item{ps_input_genetic_SD}{input file with genetic standarddeviation}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
produce output file. This function processed different functions
to extract information from the output parameter files.
}
