% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{pre_process_ew_input}
\alias{pre_process_ew_input}
\title{Pre-processing the input-parameter-file of ECOWEIGHT}
\usage{
pre_process_ew_input(
  ps_sirebreed,
  ps_dambreed,
  ps_prodsystew,
  ps_marketchannel,
  ps_path_directory2create,
  ps_input_file_literature,
  ps_input_file_par,
  ps_input_file_testedbulls,
  ps_input_file_purchasedreplacementheifers,
  ps_input_file_calving_statement,
  ps_input_file_calving,
  ps_start_date,
  ps_end_date,
  ps_input_file_progeny_flp_statement,
  ps_input_file_flp,
  ps_input_file_calf,
  ps_input_file_flp_carcass_matrix_statement,
  ps_input_file_price_cow,
  ps_input_file_price_bull,
  ps_input_file_price_heifer,
  ps_input_file_price_calf,
  ps_input_file_ped,
  pb_log,
  plogger = NULL
)
}
\arguments{
\item{ps_sirebreed}{sire breed}

\item{ps_dambreed}{dam breed}

\item{ps_prodsystew}{production system build up as option in ECOWEIGHT}

\item{ps_marketchannel}{market channel}

\item{ps_path_directory2create}{path of the directory that will be created}

\item{ps_input_file_literature}{path to file with input coming from literature for the input-parameter-file for ECOWEIGHT}

\item{ps_input_file_par}{path to file with input as parameter for each scenanario for the input-parameter-file for ECOWEIGHT}

\item{ps_input_file_testedbulls}{path to file with input for ProdSyst 1 (tested bulls) for the input-parameter-file for ECOWEIGHT}

\item{ps_input_file_purchasedreplacementheifers}{path to file with input for ProdSyst 3 (purchased replacement heifers) for the input-parameter-file for ECOWEIGHT}

\item{ps_input_file_calving_statement}{path to file with statement based on calving for the input-parameter-file for ECOWEIGHT}

\item{ps_input_file_calving}{path to file with input coming from calving for the input-parameter-file for ECOWEIGHT}

\item{ps_start_date}{setting the start of the date to filter in the data}

\item{ps_end_date}{setting the end of the date to filter in the data}

\item{ps_input_file_progeny_flp_statement}{path to file with statement based on carcass for the input-parameter-file for ECOWEIGHT}

\item{ps_input_file_flp}{path to file with input coming from carcass for the input-parameter-file for ECOWEIGHT}

\item{ps_input_file_calf}{path to file with input coming from calf price data for the input-parameter-file for ECOWEIGHT}

\item{ps_input_file_flp_carcass_matrix_statement}{path to file with statement based on carcass frequency and price for the input-parameter-file for ECOWEIGHT}

\item{ps_input_file_price_cow}{path to file with price for cows}

\item{ps_input_file_price_bull}{path to file with price for bulls}

\item{ps_input_file_price_heifer}{path to file with price for heifers}

\item{ps_input_file_price_calf}{path to file with price for calf}

\item{ps_input_file_ped}{path to pedigree file}

\item{pb_log}{indicator whether logs should be produced}

\item{plogger}{logger object}
}
\description{
The program package ECOWEIGHT (C Programs for Calculating Economic Weights in Livestock)
need input parameter files. This function processed different functions
to prepare the input parameter files.
}
